Musket moved to framework (but needs other features to be implemented)
